refactor: LottoNumberResponse renaming, 변환 로직 이동
