--- src/main/java/lotto/dto/LottoNumberDto.java
@@ -1,15 +1,0 @@
-package lotto.dto;
-
-import java.util.List;
-
-public class LottoNumberDto {
-    private List<String> numbers;
-
-    public LottoNumberDto(List<String> numbers) {
-        this.numbers = numbers;
-    }
-
-    public List<String> getNumbers() {
-        return numbers;
-    }
-}
